requirements.txt
Django~=4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-bulk==0.2.1
PyJWT==2.4.0
PyMySQL==1.0.2
loguru==0.6.0
pytest~=6.2.5
pytest-html==3.1.1
pytest-dependency==0.5.1
allure-pytest==2.9.45